backend/requirements.txt
fastapi>=0.109
uvicorn[standard]>=0.27
google-cloud-bigquery>=3.0
pandas>=2.0
pydantic>=2.0
PyYAML>=6.0
pytest>=7.4
pytest-asyncio>=0.23
